fix: split Avatar Dropdown context seperate interface,custom hooks from context
